Refinements and completed PUT /pull/:id route

+++ routes/poll.js
@@ -30,7 +30,7 @@ router.get("/poll/:id", [check.param("id").isNumeric()], async (req, res) => {
                 id: Number(record.rows[0].id),
                 name: record.rows[0].name,
                 description: record.rows[0].description,
-                options: options.rows.map(e => { return { id: Number(e.id), value: e.value, votes: Number(e.votes) } }),
+                options: options.rows.map(e => { return { value: e.value, votes: Number(e.votes) } }),
                 totalVotes: Number(record.rows[0].total_votes),
                 created: record.rows[0].created
             });
@@ -52,20 +52,21 @@ router.delete("/poll/:id", [check.param("id").isNumeric()], async (req, res) => 
         const client = await database.connect();
         try {
             await client.query("BEGIN");
-            await client.query({
-                text: "DELETE FROM poll_votes WHERE poll_id=$1",
+            const exists = (await client.query({
+                text: "SELECT FROM polls WHERE id=$1",
                 values: [req.params.id]
-            });
-            await client.query({
-                text: "DELETE FROM poll_options WHERE poll_id=$1",
-                values: [req.params.id]
-            });
+            })).rows.length === 1;
+            if (!exists) {
+                res.status(404);
+                res.send({ error: true, details: `Poll ${req.params.id} not found.`});
+                return;
+            }
             await client.query({
                 text: "DELETE FROM polls WHERE id=$1",
                 values: [req.params.id]
             });
             await client.query("COMMIT");
-            res.send({ success: true, id: Number(req.params.id) });
+            res.send({ success: true, operation: "delete", id: Number(req.params.id) });
         } catch (e) {
             await client.query("ROLLBACK");
             console.log(e);
@@ -86,8 +87,12 @@ router.put("/poll/:id", [
     check.body("description").optional().isAscii().isLength({ max: 500 }),
     check.body("options").exists().custom((options) => {
         if (!Array.isArray(options) || options.length < 2) return false;
-        for (let option of options) if (option.length > 140) return false;    
-        return true;
+            let existing = [];
+            for (let option of options) {
+                if (option.length > 140 || existing.includes(option)) return false;
+                existing.push(option);
+            }
+            return true;
     })
 ], async (req, res) => {
     res.type("application/json");
@@ -96,12 +101,38 @@ router.put("/poll/:id", [
         const client = await database.connect();
         try {
             await client.query("BEGIN");
+            const description = req.body.description ? req.body.description : "";
+            const exists = (await client.query({
+                text: "SELECT FROM polls WHERE id=$1",
+                values: [req.params.id]
+            })).rows.length === 1;
+            if (!exists) {
+                res.status(404);
+                res.send({ error: true, details: `Poll ${req.params.id} not found.`});
+                return;
+            }
             await client.query({
-                text: "UPDATE polls SET name = $1, description = $2",
-                value: [req.body.name, req.body.description, req.params.id]
+                text: "UPDATE polls SET name=$1, description=$2, modified=to_timestamp($3/1000.0) \
+                WHERE id=$4",
+                values: [req.body.name, description, Date.now(), req.params.id]
             });
+            await client.query({
+                text: "DELETE FROM poll_options WHERE NOT(value = ANY($1)) AND poll_id=$2",
+                values: [req.body.options, req.params.id]
+            });
+            const remainingOptions = (await client.query({
+                text: "SELECT value from poll_options WHERE poll_id=$1",
+                values: [req.params.id]
+            })).rows.map(row => row.value);
+            for (let option of req.body.options.filter(e => !remainingOptions.includes(String(e)))) {
+                await client.query({
+                    text: "INSERT INTO poll_options (poll_id, value, created) \
+                        VALUES ($1, $2, to_timestamp($3/1000.0))",
+                    values: [req.params.id, option, Date.now()]
+                });
+            }
             await client.query("COMMIT");
-            res.send({ success: true, poll_id: insertAndGetId.rows[0].id });
+            res.send({ success: true, operation: "update", poll_id: req.params.id});
         } catch (e) {
             await client.query("ROLLBACK");
             console.log(e);
@@ -121,7 +152,11 @@ router.post("/poll", [
         check.body("description").optional().isAscii().isLength({ max: 500 }),
         check.body("options").exists().custom((options) => {
             if (!Array.isArray(options) || options.length < 2) return false;
-            for (let option of options) if (option.length > 140) return false;    
+            let existing = [];
+            for (let option of options) {
+                if (option.length > 140 || existing.includes(option)) return false;
+                existing.push(option);
+            }
             return true;
         })
     ], async (req, res) => {
@@ -131,11 +166,13 @@ router.post("/poll", [
             const client = await database.connect();
             try {
                 await client.query("BEGIN");
+                const now = Date.now();
                 const description = req.body.description ? req.body.description : "";
                 const insertAndGetId = await client.query({
-                    text: "INSERT INTO polls (name, description, created) \
-                        VALUES ($1, $2, to_timestamp($3 / 1000.0)) RETURNING id",
-                    values: [req.body.name, description, Date.now()]
+                    text: "INSERT INTO polls (name, description, created, modified) \
+                        VALUES ($1, $2, to_timestamp($3/1000.0), to_timestamp($4/1000.0)) \
+                        RETURNING id",
+                    values: [req.body.name, description, now, now]
                 });
                 for (let option of req.body.options) {
                     await client.query({
@@ -145,7 +182,7 @@ router.post("/poll", [
                     });
                 }
                 await client.query("COMMIT");
-                res.send({ success: true, poll_id: insertAndGetId.rows[0].id });
+                res.send({ success: true, operation: "create", poll_id: insertAndGetId.rows[0].id });
             } catch (e) {
                 await client.query("ROLLBACK");
                 console.log(e);
